Feat: supporting entity tag
Add entity tags to responses (Currently a weak, dumb
entity tag implementation.
it need to be replaced by stronger one such as SHA-1
Add basic support for "If-none-Match" request headers (probably
doesn't hnadle the full syntax of that header) in order to match entity tags and return
304 Not Modified if the client
already has the same file and it is hasn't changed.